requirements.txt
pymupdf==1.23.7
PyYAML==6.0.1
tqdm==4.66.1